Reduce agent router blast radius
Split the large agent admin router into focused route groups while preserving the public router factory and adding an HTTP smoke path for capabilities.

Constraint: Issue 585 requires agent.routes.ts to be split into route-group files with units under 500 lines.

Rejected: Keep the monolithic route file | it violates the issue scope and makes future route changes high-risk.

Confidence: high

Scope-risk: moderate

Directive: Preserve route mount paths and createAgentRouter exports when touching these modules.

Tested: npx vitest run packages/memento-server/src/server/routes/agent.routes.spec.ts packages/memento-server/src/server/routes/agent-personal.routes.spec.ts; npm run test:server; npm run type-check -w memento-server; npm run lint:ts

Not-tested: npm run test:http-v2 is currently blocked by an existing @memento/core package export error outside this refactor.
